+++ studycafe/src/main/java/com/studycafe/member/controller/MemberController.java
@@ -48,6 +48,4 @@ public class MemberController {
 		
 		return "redirect:/loginForm";
 	}
-	
-
 }
