+++ studycafe/src/main/java/com/studycafe/team/teamboard/entity/TeamBoardEntity.java
@@ -1,4 +1,4 @@
-package com.studycafe.team.entity;
+package com.studycafe.team.teamboard.entity;
 
 import java.sql.Timestamp;
 
@@ -7,26 +7,35 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.Lob;
 import javax.validation.constraints.NotNull;
 
 import org.hibernate.annotations.CreationTimestamp;
 
 import lombok.Data;
 
-@Data
 @Entity
-public class TeamEntity {
+@Data
+public class TeamBoardEntity {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private long teamNumber;
+	private long teamBoardNum;
 	
 	@NotNull
-	@Column(unique=true)
-	private String teamName;
+	@Column(length = 20)
+	private String teamBoardTitle;
+	
+	@Lob
+	@NotNull
+	private String teamBoardContent;
 	
 	@NotNull
-	private String teamHead;
+	@Column(length = 20)
+	private String teamBoardWriter;
+	
+	@Column(nullable = true)
+	private String imgUrl;
 	
 	@NotNull
 	@CreationTimestamp
