+++ studycafe/src/main/java/com/studycafe/team/teamboard/repository/TeamBoardRepository.java
@@ -1,0 +1,5 @@
+package com.studycafe.team.teamboard.repository;
+
+public interface TeamBoardRepository extends JPAe{
+	// gpt api 따서 챗봇
+}
