+++ studycafe/src/main/java/com/studycafe/team/teamboard/service/TeamBoardService.java
@@ -1,0 +1,5 @@
+package com.studycafe.team.teamboard.service;
+
+public interface TeamBoardService {
+
+}
