+++ studycafe/src/main/java/com/studycafe/team/teamboard/service/TeamBoardServiceImpl.java
@@ -1,0 +1,8 @@
+package com.studycafe.team.teamboard.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class TeamBoardServiceImpl implements TeamBoardService{
+
+}
