+++ studycafe/src/main/resources/static/css/teamregis.css
@@ -75,3 +75,7 @@ form{
 	cursor: pointer;
 	color: white;
 }
+
+.select_img{
+	text-align: center;
+}
